requirements.txt
openai>=1.2.0
python-dotenv
requests
Pillow
